// src/models/card-master-data.interface.ts
export interface CardMasterInterface {
  customerName: string;
  customerMobileNo: string;
  trxId: string;
  trxType: string;
  lineId: string;
  stationId: string;
  equpGrpId: string;
  equpId: string;
  aqurId: string;
  operId: string;
  trmId: string;
  crdType: string;
  panSha: string;
  prdType: string;
  businessDt: string;
  trxDtTm: string;
}